+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.6.1] - 2021-01-11
+### Fixed
+- Environment node uses "Environment" a very often, added this to check for old enviremont textures
+- Correct default Saturation values in Node Group
+
 ## [0.0.6] - 2021-31-10
 ### Added
 - auto-transfer old environment texture images > prevents from relinking them when switching modes
@@ -41,6 +46,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
+[0.0.6.1]:https://github.com/schroef/World_Control/releases/tag/v0.0.6.1
 [0.0.6]:https://github.com/schroef/World_Control/releases/tag/v0.0.6
 [0.0.5]:https://github.com/schroef/World_Control/releases/tag/v0.0.5
 [0.0.4]:https://github.com/schroef/World_Control/releases/tag/v0.0.4